requirements.txt
requests
aiohttp
python-dotenv
aiolimiter